requirements.txt
# requirements.txt
streamlit==1.28.0
pandas==2.1.0
psycopg2-binary==2.9.7
qrcode==7.4.2
pillow==10.0.0
openpyxl==3.1.2
xlsxwriter==3.1.2
python-dotenv==1.0.0